+++ ailoop-tools/src/context.rs
@@ -540,7 +540,7 @@ mod tests {
         assert!(cloned.is_cancelled());
     }
 
-    fn tokens(input: u32, output: u32) -> Usage {
+    fn tokens(input: u64, output: u64) -> Usage {
         let mut u = Usage::default();
         u.input_tokens = input;
         u.output_tokens = output;
